Use livereload package for the auto functionality
The auto plugin seems needlessly complicated and uses ws4py which doesn't seem
to be actively maintained.  Using livereload, we can offload the burden of
maintaining the websockets related code to the developer of livereload.

It also seems to be more stable than the old code, and seems more robust
against crashing on client closing connection, etc.